[app-office/libreoffice] Resync with main tree
(Portage version: 2.1.10.9/git/Linux x86_64, RepoMan options: --force, signed Manifest commit with key B6C5F7DE)

+++ app-office/libreoffice/libreoffice-3.4.9999.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2011 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/app-office/libreoffice/libreoffice-3.4.1.ebuild,v 1.7 2011/07/18 14:22:33 scarabeus Exp $
+# $Header: /var/cvsroot/gentoo-x86/app-office/libreoffice/libreoffice-3.4.9999.ebuild,v 1.3 2011/07/31 16:03:50 mattst88 Exp $
 
 EAPI=3
 
@@ -10,58 +10,64 @@ CMAKE_REQUIRED="never"
 PYTHON_DEPEND="2"
 PYTHON_USE_WITH="threads,xml"
 
-inherit base autotools check-reqs db-use eutils fdo-mime flag-o-matic gnome2-utils java-pkg-opt-2 kde4-base multilib pax-utils prefix python toolchain-funcs
+# experimental ; release ; old
+# Usually the tarballs are moved a lot so this should make
+# everyone happy.
+DEV_URI="
+	http://dev-builds.libreoffice.org/pre-releases/src
+	http://download.documentfoundation.org/libreoffice/src
+	http://download.documentfoundation.org/libreoffice/old/src
+"
+EXT_URI="http://ooo.itc.hu/oxygenoffice/download/libreoffice"
+ADDONS_URI="http://dev-www.libreoffice.org/src/"
+
+BRANDING="${PN}-branding-gentoo-0.2.tar.xz"
+
+[[ ${PV} == *9999* ]] && SCM_ECLASS="git-2"
+inherit base autotools check-reqs eutils java-pkg-opt-2 kde4-base pax-utils prefix python multilib toolchain-funcs flag-o-matic nsplugins ${SCM_ECLASS}
+unset SCM_ECLASS
 
 DESCRIPTION="LibreOffice, a full office productivity suite."
 HOMEPAGE="http://www.libreoffice.org"
-DEV_URI="http://download.documentfoundation.org/libreoffice/src"
-EXT_URI="http://ooo.itc.hu/oxygenoffice/download/libreoffice"
-ADDONS_URI="http://dev-www.libreoffice.org/src/"
-SRC_URI="odk? ( java? ( http://tools.openoffice.org/unowinreg_prebuild/680/unowinreg.dll ) )"
+SRC_URI="branding? ( http://dev.gentooexperimental.org/~scarabeus/${BRANDING} )"
 
 # Shiny split sources with so many packages...
-MODULES="artwork base calc components extensions extras filters help
+# Bootstrap MUST be first!
+MODULES="bootstrap artwork base calc components extensions extras filters help
 impress libs-core libs-extern libs-extern-sys libs-gui postprocess sdk testing
 ure writer translations"
-# split out as bootstrap is required to be done first
-SRC_URI+=" ${DEV_URI}/${PN}-bootstrap-${PV}.tar.bz2"
-for mod in ${MODULES}; do
-	SRC_URI+=" ${DEV_URI}/${PN}-${mod}-${PV}.tar.bz2"
-done
-unset mod
+# Only release has the tarballs
+if [[ ${PV} != *9999* ]]; then
+	for i in ${DEV_URI}; do
+		for mod in ${MODULES}; do
+			SRC_URI+=" ${i}/${PN}-${mod}-${PV}.tar.bz2"
+		done
+		unset mod
+	done
+	unset i
+fi
+unset DEV_URI
 
 # addons
-ADDONS_SRC+=" ${ADDONS_URI}/48a9f787f43a09c0a9b7b00cd1fddbbf-hyphen-2.7.1.tar.gz"
-ADDONS_SRC+=" ${ADDONS_URI}/09357cc74975b01714e00c5899ea1881-pixman-0.12.0.tar.gz"
+# FIXME: actually review which one of these are used
 ADDONS_SRC+=" ${ADDONS_URI}/128cfc86ed5953e57fe0f5ae98b62c2e-libtextcat-2.2.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/17410483b5b5f267aa18b7e00b65e6e0-hsqldb_1_8_0.zip"
 ADDONS_SRC+=" ${ADDONS_URI}/1756c4fa6c616ae15973c104cd8cb256-Adobe-Core35_AFMs-314.tar.gz"
-ADDONS_SRC+=" ${ADDONS_URI}/18f577b374d60b3c760a3a3350407632-STLport-4.5.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/1f24ab1d39f4a51faf22244c94a6203f-xmlsec1-1.2.14.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/24be19595acad0a2cae931af77a0148a-LICENSE_source-9.0.0.7-bj.html"
-ADDONS_SRC+=" ${ADDONS_URI}/26b3e95ddf3d9c077c480ea45874b3b8-lp_solve_5.5.tar.gz"
-ADDONS_SRC+=" ${ADDONS_URI}/284e768eeda0e2898b0d5bf7e26a016e-raptor-1.4.18.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/2a177023f9ea8ec8bd00837605c5df1b-jakarta-tomcat-5.0.30-src.tar.gz"
-ADDONS_SRC+=" ${ADDONS_URI}/ca4870d899fd7e943ffc310a5421ad4d-liberation-fonts-ttf-1.06.0.20100721.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/35c94d2df8893241173de1d16b6034c0-swingExSrc.zip"
 ADDONS_SRC+=" ${ADDONS_URI}/35efabc239af896dfb79be7ebdd6e6b9-gentiumbasic-fonts-1.10.zip"
 ADDONS_SRC+=" ${ADDONS_URI}/39bb3fcea1514f1369fcfc87542390fd-sacjava-1.3.zip"
-ADDONS_SRC+=" ${ADDONS_URI}/3ade8cfe7e59ca8e65052644fed9fca4-epm-3.7.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/48470d662650c3c074e1c3fabbc67bbd-README_source-9.0.0.7-bj.txt"
 ADDONS_SRC+=" ${ADDONS_URI}/4a660ce8466c9df01f19036435425c3a-glibc-2.1.3-stub.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/599dc4cc65a07ee868cf92a667a913d2-xpdf-3.02.tar.gz"
-ADDONS_SRC+=" ${ADDONS_URI}/7376930b0d3f3d77a685d94c4a3acda8-STLport-4.5-0119.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/798b2ffdc8bcfe7bca2cf92b62caf685-rhino1_5R5.zip"
 ADDONS_SRC+=" ${ADDONS_URI}/8294d6c42e3553229af9934c5c0ed997-stax-api-1.0-2-sources.jar"
 ADDONS_SRC+=" ${ADDONS_URI}/a7983f859eafb2677d7ff386a023bc40-xsltml_2.1.2.zip"
 ADDONS_SRC+=" ${ADDONS_URI}/ada24d37d8d638b3d8a9985e80bc2978-source-9.0.0.7-bj.zip"
-ADDONS_SRC+=" ${ADDONS_URI}/c441926f3a552ed3e5b274b62e86af16-STLport-4.0.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/d4c4d91ab3a8e52a2e69d48d34ef4df4-core.zip"
-ADDONS_SRC+=" ${ADDONS_URI}/e0707ff896045731ff99e99799606441-README_db-4.7.25.NC-custom.txt"
-ADDONS_SRC+=" ${ADDONS_URI}/fca8706f2c4619e2fa3f8f42f8fc1e9d-rasqal-0.9.16.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/fdb27bfe2dbe2e7b57ae194d9bf36bab-SampleICC-1.3.2.tar.gz"
-ADDONS_SRC+=" ${ADDONS_URI}/37282537d0ed1a087b1c8f050dc812d9-dejavu-fonts-ttf-2.32.zip"
-ADDONS_SRC+=" ${ADDONS_URI}/067201ea8b126597670b5eff72e1f66c-mythes-1.2.0.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/3404ab6b1792ae5f16bbd603bd1e1d03-libformula-1.1.7.zip"
 ADDONS_SRC+=" ${ADDONS_URI}/3bdf40c0d199af31923e900d082ca2dd-libfonts-1.1.6.zip"
 ADDONS_SRC+=" ${ADDONS_URI}/8ce2fcd72becf06c41f7201d15373ed9-librepository-1.1.6.zip"
@@ -76,23 +82,10 @@ ADDONS_SRC+=" ${ADDONS_URI}/47e1edaa44269bc537ae8cabebb0f638-JLanguageTool-1.0.0
 ADDONS_SRC+=" ${ADDONS_URI}/90401bca927835b6fbae4a707ed187c8-nlpsolver-0.9.tar.bz2"
 ADDONS_SRC+=" ${ADDONS_URI}/0f63ee487fda8f21fafa767b3c447ac9-ixion-0.2.0.tar.gz"
 ADDONS_SRC+=" ${ADDONS_URI}/71474203939fafbe271e1263e61d083e-nss-3.12.8-with-nspr-4.8.6.tar.gz"
-ADDONS_SRC+=" ${ADDONS_URI}/7a0dcb3fe1e8c7229ab4fb868b7325e6-mdds_0.5.2.tar.bz2"
-ADDONS_SRC+=" ${ADDONS_URI}/0625a7d661f899a8ce263fc8a9879108-graphite2-0.9.2.tgz"
 ADDONS_SRC+=" http://download.go-oo.org/extern/185d60944ea767075d27247c3162b3bc-unowinreg.dll"
 ADDONS_SRC+=" http://download.go-oo.org/extern/b4cae0700aa1c2aef7eb7f345365e6f1-translate-toolkit-1.8.1.tar.bz2"
 ADDONS_SRC+=" http://www.numbertext.org/linux/881af2b7dca9b8259abbca00bbbc004d-LinLibertineG-20110101.zip"
 SRC_URI+=" ${ADDONS_SRC}"
-
-# translations
-LANGUAGES="af ar as ast be bg bn bo br brx bs ca ca_XV cs cy da de dgo dz el
-en en_GB en_ZA eo es et eu fa fi fr ga gl gu he hi hr hu id is it ja ka kk km
-kn kok ko ks ku lo lt lv mai mk ml mn mni mr my nb ne nl nn nr nso oc or
-pa_IN pl pt pt_BR ro ru rw sat sd sh sk sl sq sr ss st sv sw_TZ ta te tg
-th tn tr ts ug uk uz ve vi xh zh_CN zh_TW zu"
-for X in ${LANGUAGES} ; do
-	IUSE+=" linguas_${X}"
-done
-unset X
 
 # intersection of available linguas and app-dicts/myspell-* dictionaries
 SPELL_DIRS="af bg ca cs cy da de el en eo es et fr ga gl he hr hu it ku lt mk nb
@@ -116,11 +109,23 @@ unset ADDONS_URI
 unset EXT_URI
 unset ADDONS_SRC
 
-IUSE+=" binfilter cups custom-cflags dbus debug eds gnome gstreamer
-gtk kde ldap mysql nsplugin odk offlinehelp opengl pch python templates webdav"
+IUSE="binfilter +branding cups custom-cflags dbus debug eds gnome graphite
+gstreamer gtk kde ldap mysql nsplugin odk offlinehelp opengl python templates
+test +vba webdav"
 LICENSE="LGPL-3"
 SLOT="0"
-KEYWORDS="~amd64 ~ppc ~x86 ~amd64-linux ~x86-linux"
+[[ ${PV} == *9999* ]] || KEYWORDS="~amd64 ~ppc ~x86 ~amd64-linux ~x86-linux"
+
+# translations
+LANGUAGES="af ar as ast be bg bn bo br brx bs ca ca_XV cs cy da de dgo dz el
+en en_GB en_ZA eo es et eu fa fi fr ga gl gu he hi hr hu id is it ja ka kk km
+kn kok ko ks ku lo lt lv mai mk ml mn mni mr my nb ne nl nn nr nso oc or
+pa_IN pl pt pt_BR ro ru rw sat sd sh sk sl sq sr ss st sv sw_TZ ta te tg
+th tn tr ts ug uk uz ve vi xh zh_CN zh_TW zu"
+for X in ${LANGUAGES} ; do
+	IUSE+=" linguas_${X}"
+done
+unset X
 
 COMMON_DEPEND="
 	app-arch/zip
@@ -133,6 +138,7 @@ COMMON_DEPEND="
 	dev-db/unixODBC
 	dev-libs/expat
 	>=dev-libs/glib-2.18
+	>=dev-libs/hyphen-2.7.1
 	>=dev-libs/icu-4.0
 	>=dev-lang/perl-5.0
 	>=dev-libs/openssl-0.9.8g
@@ -142,6 +148,7 @@ COMMON_DEPEND="
 	>=media-libs/vigra-1.4
 	media-libs/libpng
 	media-libs/libwpg:0.2
+	sci-mathematics/lpsolve
 	>=sys-libs/db-4.8
 	virtual/jpeg
 	>=x11-libs/cairo-1.0.2
@@ -157,6 +164,7 @@ COMMON_DEPEND="
 		gnome-base/gconf:2
 	)
 	gtk? ( >=x11-libs/gtk+-2.10:2 )
+	graphite? ( media-gfx/graphite2 )
 	gstreamer? (
 		>=media-libs/gstreamer-0.10
 		>=media-libs/gst-plugins-base-0.10
@@ -194,6 +202,7 @@ DEPEND="${COMMON_DEPEND}
 	dev-util/cppunit
 	>=dev-util/gperf-3
 	dev-util/intltool
+	dev-util/mdds
 	dev-util/pkgconfig
 	media-gfx/sane-backends
 	>=net-misc/curl-7.12
@@ -203,7 +212,6 @@ DEPEND="${COMMON_DEPEND}
 	sys-devel/flex
 	sys-libs/zlib
 	x11-libs/libXtst
-	x11-proto/printproto
 	x11-proto/randrproto
 	x11-proto/xextproto
 	x11-proto/xineramaproto
@@ -211,7 +219,7 @@ DEPEND="${COMMON_DEPEND}
 	java? (
 		=virtual/jdk-1.6*
 		>=dev-java/ant-core-1.7
-		dev-java/junit:4
+		test? ( dev-java/junit:4 )
 	)
 "
 
@@ -220,12 +228,20 @@ PATCHES=(
 	"${FILESDIR}/${PN}-libdb5-fix-check.diff"
 	"${FILESDIR}/${PN}-3.4.1-salfix.diff"
 	"${FILESDIR}/sdext-presenter.diff"
+	"${FILESDIR}/${PN}-svx.patch"
+	"${FILESDIR}/${PN}-vbaobj-visibility-fix.patch"
+	"${FILESDIR}/${PN}-solenv-build-crash.patch"
+	"${FILESDIR}/${PN}-as-needed-gtk.patch"
+	"${FILESDIR}/${PN}-translate-toolkit-parallel-solenv.patch"
+	"${FILESDIR}/${PN}-gbuild-use-cxxflags.patch"
+	"${FILESDIR}/${PN}-installed-files-permissions.patch"
 )
 
 # Uncoment me when updating to eapi4
 # REQUIRED_USE="
 #	|| ( gtk gnome kde )
 #	gnome? ( gtk )
+#	nsplugin? ( gtk )
 #"
 
 S="${WORKDIR}/${PN}-bootstrap-${PV}"
@@ -236,6 +252,11 @@ pkg_setup() {
 
 	python_set_active_version 2
 	python_pkg_setup
+
+	if [[ $(gcc-major-version) -lt 4 ]]; then
+		eerror "Compilation with gcc older than 4.0 is not supported"
+		die "Too old gcc found."
+	fi
 
 	if use custom-cflags; then
 		ewarn "You are using custom CFLAGS, which is NOT supported and can cause"
@@ -264,28 +285,47 @@ pkg_setup() {
 		ewarn
 	fi
 
+	ewarn "Libreoffice compilation often fails on parallel issues"
+	ewarn "but the slowdown by enforcing MAKEOPTS=-j1 is too huge."
+	ewarn "If you encounter errors try yourself to disable parallel build."
+
 	# Check if we have enough RAM and free diskspace to build this beast
-	CHECKREQS_MEMORY="512"
-	use debug && CHECKREQS_DISK_BUILD="12288" || CHECKREQS_DISK_BUILD="7144"
+	CHECKREQS_MEMORY="1024"
+	use debug && CHECKREQS_DISK_BUILD="15360" || CHECKREQS_DISK_BUILD="9216"
 	check_reqs
 }
 
 src_unpack() {
-	local mod dest tmplfile tmplname
+	local mod dest tmplfile tmplname mypv
 
-	#first the bootstrap files
-	unpack "${PN}-bootstrap-${PV}.tar.bz2"
+	if use branding; then
+		unpack "${BRANDING}"
+	fi
 
-	# and then all the separate modules
-	for mod in ${MODULES}; do
-		unpack "${PN}-${mod}-${PV}.tar.bz2"
-		mv -n "${WORKDIR}/${PN}-${mod}-${PV}"/* "${S}"
-		# punt the empty dirs; it is annoying during debuging :)
-		rm -rf "${WORKDIR}/${PN}-${mod}-${PV}"
-	done
-
-	# don't forget the wrapper...
-	cp "${FILESDIR}"/wrapper.in "${T}"
+	if [[ ${PV} != *9999* ]]; then
+		for mod in ${MODULES}; do
+			unpack "${PN}-${mod}-${PV}.tar.bz2"
+			if [[ ${mod} != bootstrap ]]; then
+				mv -n "${WORKDIR}/${PN}-${mod}-${PV}"/* "${S}"
+				rm -rf "${WORKDIR}/${PN}-${mod}-${PV}"
+			fi
+		done
+	else
+		for mod in ${MODULES}; do
+			mypv=${PV/.9999}
+			[[ ${mypv} != ${PV} ]] && EGIT_BRANCH="${PN}-${mypv/./-}"
+			EGIT_PROJECT="${PN}/${mod}"
+			EGIT_SOURCEDIR="${WORKDIR}/${PN}-${mod}-${PV}"
+			EGIT_REPO_URI="git://anongit.freedesktop.org/${PN}/${mod}"
+			EGIT_NOUNPACK="true"
+			git-2_src_unpack
+			if [[ ${mod} != bootstrap ]]; then
+				mv -n "${WORKDIR}/${PN}-${mod}-${PV}"/* "${S}"
+				rm -rf "${WORKDIR}/${PN}-${mod}-${PV}"
+			fi
+		done
+		unset EGIT_PROJECT EGIT_SOURCEDIR EGIT_REPO_URI EGIT_BRANCH
+	fi
 
 	# copy extension templates; o what fun ...
 	if use templates; then
@@ -307,8 +347,6 @@ src_unpack() {
 }
 
 src_prepare() {
-	eprefixify "${T}"/wrapper.in
-
 	strip-linguas ${LANGUAGES}
 	LINGUAS_OOO=$(echo ${LINGUAS} | sed -e 's/\ben\b/en_US/;s/_/-/g')
 
@@ -322,12 +360,6 @@ src_prepare() {
 	# silent miscompiles; LO/OOo adds -O2/1/0 where appropriate
 	filter-flags "-O*"
 
-	if [[ $(gcc-major-version) -lt 4 ]]; then
-		filter-flags "-fstack-protector"
-		filter-flags "-fstack-protector-all"
-		replace-flags "-fomit-frame-pointer" "-momit-leaf-frame-pointer"
-	fi
-
 	base_src_prepare
 	eautoreconf
 }
@@ -336,14 +368,11 @@ src_configure() {
 	local java_opts
 	local internal_libs
 	local extensions
-	local themes="default"
-	local jobs=$(echo "${MAKEOPTS}" | sed -e "s/.*-j\([0-9]\+\).*/\1/")
+	local themes="crystal"
+	local jbs=$(sed -ne 's/.*\(-j[[:space:]]*\|--jobs=\)\([[:digit:]]\+\).*/\2/;T;p' <<< "${MAKEOPTS}")
 
-	# ensure that qt4 and kdedir are properly located
-	if use kde; then
-		export KDE4DIR="${KDEDIR}"
-		export QT4LIB="${EPREFIX}/usr/$(get_libdir)/qt4"
-	fi
+	# recheck that there is some value in jobs
+	[[ -z ${jbs} ]] && jbs="1"
 
 	# expand themes we are going to build based on DE useflags
 	use gnome && themes+=" tango"
@@ -356,14 +385,10 @@ src_configure() {
 		--enable-ext-presenter-minimizer
 	"
 
-	# Things that do not have gentoo packages
 	# hsqldb: requires just 1.8.0 not 1.8.1 which we don't ship at all
-	# we should use in-system dmake: so far fails
+	# dmake: not worth of splitting out
 	internal_libs+="
-		--without-system-altlinuxhyph
 		--without-system-hsqldb
-		--without-system-lpsolve
-		--without-system-mdds
 	"
 
 	# When building without java some things needs to be done
@@ -373,24 +398,32 @@ src_configure() {
 			--without-system-beanshell
 			--without-system-lucene
 			--without-system-saxon
+			--without-junit
 		"
 	else
 		java_opts="
-			--with-ant-home=${ANT_HOME}
+			--with-ant-home="${ANT_HOME}"
 			--with-jdk-home=$(java-config --jdk-home 2>/dev/null)
 			--with-java-target-version=$(java-pkg_get-target)
-			--with-jvm-path=${EPREFIX}/usr/$(get_libdir)/
+			--with-jvm-path="${EPREFIX}/usr/$(get_libdir)/"
 			--with-beanshell-jar=$(java-pkg_getjar bsh bsh.jar)
 			--with-lucene-core-jar=$(java-pkg_getjar lucene-2.9 lucene-core.jar)
 			--with-lucene-analyzers-jar=$(java-pkg_getjar lucene-analyzers-2.3 lucene-analyzers.jar)
 			--with-saxon-jar=$(java-pkg_getjar saxon saxon8.jar)
-			--with-junit=$(java-pkg_getjar junit-4 junit.jar)
 		"
+		if use test; then
+			java_opts+=" --with-junit=$(java-pkg_getjar junit-4 junit.jar)"
+		else
+			java_opts+=" --without-junit"
+		fi
 	fi
 
-	# TODO: create gentoo branding on the about/intro screens
-	# --with-about-bitmap="${FILESDIR}/gentoo-about.png"
-	# --with-intro-bitmap="${FILESDIR}/gentoo-intro.png"
+	if use branding; then
+		extensions+="
+			--with-about-bitmap="${WORKDIR}/branding-about.png"
+			--with-intro-bitmap="${WORKDIR}/branding-intro.png"
+		"
+	fi
 
 	# system headers/libs/...: enforce using system packages
 	#   only expections are mozilla and odbc/sane/xrender-header(s).
@@ -398,12 +431,12 @@ src_configure() {
 	# --enable-unix-qstart-libpng: use libpng splashscreen that is faster
 	# --disable-broffice: do not use brazillian brand just be uniform
 	# --enable-cairo: ensure that cairo is always required
-	# --disable-graphite: no package in gentoo
 	# --enable-*-link: link to the library rather than just dlopen on runtime
 	# --disable-fetch-external: prevent dowloading during compile phase
 	# --disable-gnome-vfs: old gnome virtual fs support
 	# --disable-kdeab: kde3 adressbook
 	# --disable-kde: kde3 support
+	# --disable-pch: precompiled headers cause build crashes
 	# --disable-rpath: relative runtime path is not desired
 	# --disable-static-gtk: ensure that gtk is linked dynamically
 	# --disable-zenity: disable build icon
@@ -431,12 +464,13 @@ src_configure() {
 		--disable-epm \
 		--disable-fetch-external \
 		--disable-gnome-vfs \
-		--disable-graphite \
 		--disable-kdeab \
 		--disable-kde \
 		--disable-online-update \
+		--disable-pch \
 		--disable-rpath \
 		--disable-static-gtk \
+		--disable-strip-solver \
 		--disable-zenity \
 		--with-alloc=system \
 		--with-build-version="Gentoo official package" \
@@ -446,8 +480,8 @@ src_configure() {
 		--with-external-thes-dir="${EPREFIX}/usr/share/myspell" \
 		--with-external-tar="${DISTDIR}" \
 		--with-lang="${LINGUAS_OOO}" \
-		--with-max-jobs=${jobs} \
-		--with-num-cpus=${jobs} \
+		--with-max-jobs=${jbs} \
+		--with-num-cpus=1 \
 		--with-theme="${themes}" \
 		--with-unix-wrapper=libreoffice \
 		--with-vendor="Gentoo Foundation" \
@@ -460,10 +494,12 @@ src_configure() {
 		$(use_enable binfilter) \
 		$(use_enable cups) \
 		$(use_enable dbus) \
+		$(use_enable debug crashdump) \
 		$(use_enable eds evolution2) \
 		$(use_enable gnome gconf) \
 		$(use_enable gnome gio) \
 		$(use_enable gnome lockdown) \
+		$(use_enable graphite) \
 		$(use_enable gstreamer) \
 		$(use_enable gtk) \
 		$(use_enable gtk systray) \
@@ -474,72 +510,40 @@ src_configure() {
 		$(use_enable nsplugin mozilla) \
 		$(use_enable odk) \
 		$(use_enable opengl) \
-		$(use_enable pch) \
 		$(use_enable python) \
 		$(use_enable python ext-scripting-python) \
+		$(use_enable vba) \
+		$(use_enable vba activex-component) \
 		$(use_enable webdav neon) \
 		$(use_with java) \
-		$(use_with java junit) \
 		$(use_with ldap openldap) \
 		$(use_with mysql system-mysql-cppconn) \
 		$(use_with nsplugin system-mozilla libxul) \
 		$(use_with offlinehelp helppack-integration) \
 		$(use_with templates sun-templates) \
 		${internal_libs} \
-		${java_opts}
+		${java_opts} \
+		${extensions}
 }
 
 src_compile() {
-	emake || die
+	# this is not a proper make script and the jobs are passed during configure
+	make || die
 }
 
 src_install() {
-	local SIZE desk app
+	# This is not Makefile so no buildserver
+	make DESTDIR="${D}" distro-pack-install || die
 
-	export PYTHONPATH=""
+	# symlink the plugin to system location
+	if use nsplugin; then
+		inst_plugin /usr/$(get_libdir)/libreoffice/program/libnpsoplugin.so
+	fi
 
-	emake DESTDIR="${D}" install || die
-
-	# Fix the permissions for security reasons
-	use prefix || chown -RP root:0 "${ED}"
-
-	# Desktop files
-	for i in *; do
-		mv ${i}.desktop ${PN}-${i}.desktop
-	done
-	sed -i \
-		-e s/libreoffice3.4/${PN}/g \
-		-e s/libreoffice34/${PN}/g \
-		"${ED}"/usr/$(get_libdir)/${PN}/share/xdg/*.desktop || die111
-	use java || rm "${ED}"/usr/$(get_libdir)/${PN}/share/xdg/javafilter.desktop
-	pushd "${ED}"/usr/$(get_libdir)/${PN}/share/xdg/ > /dev/null
-	popd > /dev/null
-	domenu "${ED}"/usr/$(get_libdir)/${PN}/share/xdg/*.desktop
-
-	# install icons
-	insinto /usr/share/icons/
-	doins -r "${S}"/sysui/desktop/icons/hicolor
-
-	# app icon names are too generic, have to make them unique
-	for SIZE in 16 32 48 128 ; do
-		cd "${ED}"/usr/share/icons/hicolor/${SIZE}x${SIZE}/apps
-		for app in base calc draw impress main math startcenter writer ; do
-			mv ${app}.png ${PN}-${app}.png || die
-		done
-	done
-
-	# install mime package
-	dodir /usr/share/mime/packages
-	cp sysui/*.pro/misc/${PN}/openoffice.org.xml \
-		"${ED}"/usr/share/mime/packages/${PN}.xml
-
-	# Install wrapper script
-	sed -i -e s/LIBDIR/$(get_libdir)/g "${T}/wrapper.in" || die
-	newbin "${T}/wrapper.in" ${PN} || die
-
-	# Cleanup after playing
-	rm "${ED}"/gid_Module_*
-
+	if use branding; then
+		insinto /usr/$(get_libdir)/${PN}/program
+		newins "${WORKDIR}/branding-sofficerc" sofficerc || die
+	fi
 }
 
 pkg_preinst() {
